migrated denoising to stop recording
